selection state is now free of blueprint dependencies
