+++ crates/re_viewer/src/ui/viewport.rs
@@ -8,7 +8,10 @@ use itertools::Itertools as _;
 use re_data_store::EntityPath;
 
 use crate::{
-    misc::{space_info::SpaceInfoCollection, Item, SpaceViewHighlights, ViewerContext},
+    misc::{
+        highlights_for_space_view, space_info::SpaceInfoCollection, Item, SpaceViewHighlights,
+        ViewerContext,
+    },
     ui::space_view_heuristics::default_created_space_views,
 };
 
@@ -662,10 +665,8 @@ impl<'a, 'b> egui_dock::TabViewer for TabViewer<'a, 'b> {
     fn ui(&mut self, ui: &mut egui::Ui, space_view_id: &mut Self::Tab) {
         crate::profile_function!();
 
-        let highlights = self
-            .ctx
-            .selection_state()
-            .highlights_for_space_view(*space_view_id, self.space_views);
+        let highlights =
+            highlights_for_space_view(self.ctx.selection_state(), *space_view_id, self.space_views);
         let space_view = self
             .space_views
             .get_mut(space_view_id)
